added deletion dependency management

+++ tpkg.c
@@ -3,8 +3,9 @@
 #include <stdlib.h>
 #include <string.h>
 #include <sys/stat.h>
+#include <dirent.h>	// DIR struct
 
-#define ROOT "/home/richard/root"
+#define ROOT "/home/richard/root"	// fake root directory to operate on (for safety)
 #define URL "https://raw.githubusercontent.com/Unturned3/unturned3.github.io/master/repo"
 // #define URL "127.0.0.1:8080"
 
@@ -22,36 +23,51 @@ enum wgetErr {
 mode_t dirPerm = S_IRWXU | S_IRGRP | S_IXGRP | S_IROTH | S_IXOTH;
 
 void disp_usage() { printf("error: incorrect/missing arguments\n"); }
-const char* optString = "ir";
+const char* optString = "ir";	// accepted options for tpkg
 
 int listTop = 0;
 char pkgList[32][128];
 
-int installed(char *name)
+int installed(char *name)	// check if a package is installed already
 {
 	char cmd[256];
 	sprintf(cmd, "usr/bin/tpkg_DB/%s", name);
 	if(access(cmd, F_OK) == 0)
-		return 1;
+		return 1;	// file exists
 	return 0;
 }
 
 int chdb(int action, char *name)
 {
 	char cmd[256];
-	if(action == 1)
+	if(action == 1)	// creates an entry in tpkg_DB
 		sprintf(cmd, "mkdir usr/bin/tpkg_DB/%s &> /dev/null", name);
-	if(action == -1)
-		sprintf(cmd, "rm -r usr/bin/tpkg_DB/%s &> /dev/null", name);
-	int s = WEXITSTATUS(system(cmd));
+	if(action == -1) {	// removes an entry in tpkg_DB
+		sprintf(cmd, "usr/bin/tpkg_DB/%s", name);
+		int n = 0;
+		DIR *d = opendir(cmd);
+		if(d == NULL)
+			return 1;	// directory does not exist (pkg not installed)
+		while(readdir(d) != NULL) {
+			n++;	// count how many files are present inside directory
+			if(n > 2)
+				break;
+		}
+		closedir(d);
+		if(n <= 2)	// directory is empty
+			sprintf(cmd, "rmdir usr/bin/tpkg_DB/%s &> /dev/null", name);
+		else
+			return 2;	// directory is NOT empty (pkg required by others)
+	}
+	int s = WEXITSTATUS(system(cmd));	// rmdir & mkdir returns 1 if failed
 	return s;
 }
 
-int install(char name[128])
+int install(char name[128], char rqb[128])
 {
-	printf("Installing package \"%s\"\n", name);
+	// printf("Installing package \"%s\"\n", name);
 	if(chdb(1, name) == 1) {
-		printf("  already installed\n", name);
+		printf("\"%s\" is already installed\n", name);
 		return 0;
 	}
 	
@@ -59,9 +75,9 @@ int install(char name[128])
 	int stat1 = 0, stat2 = 0;
 
 	// download the pkg and its .dep file
-	sprintf(cmd, "wget -cq -P usr/bin %s/%s", URL, name);
+	sprintf(cmd, "wget -c -P usr/bin %s/%s", URL, name);
 	stat1 = WEXITSTATUS(system(cmd));
-	sprintf(cmd, "wget -cq -P usr/bin/tpkg_DB %s/%s.dep", URL, name);
+	sprintf(cmd, "wget -c -P usr/bin/tpkg_DB %s/%s.dep", URL, name);
 	stat2 = WEXITSTATUS(system(cmd));
 	if(stat1 == wget_OK && stat2 == wget_OK) {
 		// modify file permission
@@ -69,14 +85,22 @@ int install(char name[128])
 		system(cmd);
 	} else if (stat1 == wget_server_err || stat2 == wget_server_err) {
 		// server returned error, such as 404, etc.
-		printf("  wget server side error. Does \"%s\" exist?\n", name);
+		printf("wget error. Does \"%s\" exist?\n", name);
 		goto PKG_FAILED;
 	} else {
-		printf("\n  wget unknown error, aborting...\n");
+		printf("\nwget unknown error, aborting...\n");
 		goto PKG_FAILED;
 	}
 
-	printf("  Resolving dependencies...\n");
+	// add package "rqb" into current package's tpkg_DB directory
+	// meaning: package "rqb" requires current package as a dependency
+	// "null": nothing depends on the current packge, so skip this
+	if(strcmp(rqb, "null") != 0) {
+		sprintf(cmd, "touch usr/bin/tpkg_DB/%s/%s", name, rqb);
+		system(cmd);
+	}
+
+	printf("Resolving dependencies...\n");
 
 	char depaddr[256];
 	int depSize = 0;
@@ -87,13 +111,14 @@ int install(char name[128])
 	FILE *depfp = fopen(depaddr, "r");
 	while(fscanf(depfp, "%128s", depList[depSize]) != EOF)
 		depSize++;
+	fclose(depfp);
 	
 	// download dependencies
 	for(int i=0; i<depSize; i++) {
 		if(installed(depList[i]))
 			continue;
-		if(install(depList[i]) == -1) {
-			fprintf(stderr, "  error fetching dependency \"%s\"\n", depList[i]);
+		if(install(depList[i], name) == -1) {
+			fprintf(stderr, "Error fetching dependency \"%s\"\n", depList[i]);
 			goto PKG_FAILED;
 		}
 	}
@@ -103,7 +128,7 @@ PKG_FAILED:
 	sprintf(cmd, "rm -r usr/bin/%s &> /dev/null", name);
 	system(cmd);	// remove pkg in usr/bin
 	chdb(-1, name);	// remove record in tpkg_DB
-	chdb(-1, strcat(name, ".dep"));
+	chdb(-1, strcat(name, ".dep"));	// remove its corresponding .dep file
 	return -1; 
 }
 
@@ -111,23 +136,20 @@ int main(int argc, char** argv)
 {
 	if(argc == 1) {
 		disp_usage();
-		return 0;	// exit if no options provided
+		return 1;	// exit if no options provided
 	}
 
 	// is tpkg running as root?
 	if(getuid() != 0) {
 		fprintf(stderr, "tpkg must run as root!\n");
-		return 0;
+		return 1;
 	}
 
 	// change the working directory to ROOT
 	if(chdir(ROOT) == -1) {
 		fprintf(stderr, "cannot set working directory. Does ROOT exist?\n");
-		return 0;
+		return 1;
 	}
-
-	// reset umask for correct permissions
-	// umask(0);
 
 	// create usr/bin/tpkg_DB if it doesn't exist
 	if(access("usr/bin/tpkg_DB", F_OK) != 0) {
@@ -170,7 +192,7 @@ int main(int argc, char** argv)
 	// install the requested packages
 	if(opt.ins) {
 		for(int i=0; i<listTop; i++) {
-			int status = install(pkgList[i]);	// DFS-install pkg
+			int status = install(pkgList[i], "null");	// DFS-install pkg
 			if(status == -1) {
 				printf("error occured. aborting...\n");
 				return 0;
@@ -182,14 +204,21 @@ int main(int argc, char** argv)
 	if(opt.rm) {
 		for(int i=0; i<listTop; i++) {
 			int stat = 0;
-			char cmd[256];
+			char cmd[256] = "usr/bin/tpkg_DB/";
 			stat = chdb(-1, pkgList[i]);	// remove record in tpkg_DB
 			if(stat == 1) {
 				fprintf(stderr, "package \"%s\" is not installed...\n", pkgList[i]);
 				continue;
 			}
+			if(stat == 2) {
+				fprintf(stderr, "cannot remove package \"%s\". Required by: ", pkgList[i]);
+				sprintf(cmd, "ls usr/bin/tpkg_DB/%s", pkgList[i]);
+				system(cmd);
+				continue;
+			}
 			sprintf(cmd, "rm -r usr/bin/%s &> /dev/null", pkgList[i]);
 			system(cmd);	// remove pkg installed in usr/bin
+			// remove its dependency requirement from other packages
 			chdb(-1, strcat(pkgList[i], ".dep"));	// remove .dep file in tpkg_DB
 		}
 	}
